+++ ddtrace/contrib/internal/tornado/constants.py
@@ -1,0 +1,7 @@
+"""
+This module defines Tornado settings that are shared between
+integration modules.
+"""
+CONFIG_KEY = "datadog_trace"
+REQUEST_SPAN_KEY = "__datadog_request_span"
+FUTURE_SPAN_KEY = "__datadog_future_span"
