+++ ddtrace/contrib/unittest/__init__.py
@@ -35,13 +35,18 @@ Global Configuration
    Default: ``True``
 """
 from ...internal.utils.importlib import require_modules
-from .patch import get_version
-from .patch import patch
-from .patch import unpatch
 
 
 required_modules = ["unittest"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
+        # Required to allow users to import from `ddtrace.contrib.unittest.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.unittest.patch import get_version
+        from ..internal.unittest.patch import patch
+        from ..internal.unittest.patch import unpatch
+
         __all__ = ["patch", "unpatch", "get_version"]
