+++ ddtrace/contrib/urllib3/__init__.py
@@ -51,13 +51,18 @@ Global Configuration
    Default: ``False``
 """
 from ...internal.utils.importlib import require_modules
-from .patch import get_version
-from .patch import patch
-from .patch import unpatch
 
 
 required_modules = ["urllib3"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
+        # Required to allow users to import from `ddtrace.contrib.urllib3.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.urllib3.patch import get_version
+        from ..internal.urllib3.patch import patch
+        from ..internal.urllib3.patch import unpatch
+
         __all__ = ["patch", "unpatch", "get_version"]
