+++ ddtrace/contrib/wsgi/__init__.py
@@ -36,8 +36,8 @@ Global Configuration
 :ref:`All HTTP tags <http-tagging>` are supported for this integration.
 
 """
-from .wsgi import DDWSGIMiddleware
-from .wsgi import get_version
+from ..wsgi import DDWSGIMiddleware
+from ..wsgi import get_version
 
 
 __all__ = ["DDWSGIMiddleware", "get_version"]
