+++ ddtrace/contrib/yaaredis/__init__.py
@@ -73,7 +73,11 @@ required_modules = ["yaaredis", "yaaredis.client"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.yaaredis.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.yaaredis.patch import get_version
+        from ..internal.yaaredis.patch import patch
 
         __all__ = ["patch", "get_version"]
